http: remove unnecessary cast
Fixes the clippy warning:

warning: casting to the same type is unnecessary (`usize` -> `usize`)
   --> proxmox-http/src/websocket/mod.rs:446:40
    |
446 |             mask.copy_from_slice(&data[mask_offset as usize..payload_offset as usize]);
    |                                        ^^^^^^^^^^^^^^^^^^^^ help: try: `mask_offset`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
    = note: `#[warn(clippy::unnecessary_cast)]` on by default

+++ proxmox-http/src/websocket/mod.rs
@@ -441,7 +441,7 @@ impl FrameHeader {
                 return Ok(None);
             }
             let mut mask = [0u8; 4];
-            mask.copy_from_slice(&data[mask_offset as usize..payload_offset as usize]);
+            mask.copy_from_slice(&data[mask_offset..payload_offset as usize]);
             Some(mask)
         } else {
             None
